vmeio: remove unsightly casts
Several casts are spread around the code. Most of them we can get
rid of, increasing
* clarity
* safety
* cleanliness

In particular, byte read/write operations were inconsistently
performed by just accessing IO mapped addresses, which is incorrect
(ioread8/iowrite8 should be used instead, not only for endianness
but for memory barrier adjunction). Casts to and from void* to (T*)
are also superfluous and confusing

More casts remain that can be deleted when a saner structure for
the ioctl method is put in place