vmeio: change module contexts by device structures
A module is defined to be an array of memory-mapped areas
and to possibly originate an interrupt. A vmeio_device is
defined accordingly.

This addresses several design flaws. First, struct vmeio_get_window_s
contains information that does NOT belong to a window, like data of
interrupt sources or whether DMA should be globally used or not; this has been
factored out to vmeio_device. Second, statically declaring maps by
giving 8 fields named XXX1 and XXX2 gives rise to reams of lines of
repeated code. See for example what we need to do in order to
dispatch on the window number (around line #640 or #670).
This is verbose and ugly beyond measure: it could be tolerated
given that two maps are defined statically by command-line,
but we need not pollute other code innecessarily because of this.

Another failure in the design is that maps are, well, mapped all in a
row. We cannot map/unmap selectively; worse than that, SET_DEVICE
unmaps everything AND de-registers the isr. This has to be solved
and this is the first step to it.