vmeio: leave only base addresses and interrupt vec as vector parameters
There is no point in having address modifiers, data widths and
lengths of each window vary from one module to another if they
are the same class of hardware. The only parameters that
vary by lun are base address and interrupt vector. This vastly
shortens the code and suppresses a (mis)feature that was not
being used anyway

Signed-off-by: Juan David Gonzalez Cobas <dcobas@cern.ch>