vmeio: dump user_driver_parameters.h
This header included several unrelated definitions, thus making
for unclear dependencies. First of all, DRV_NAME is scattered all
over the place, and needs to be seen by vmeio.c and the support library
(really?). Second, the DRV_SYMB hack is not needed in the driver, so
there's no point in having it there. Third, the DRV_MAX_DEVICES does
belong to it, so it should go to vmeio.h

This commit sorts out the mess. Currently, "vmeio" is set as the
default driver name, to be overriden by a global compile-time switch
-DDRIVER_NAME="something"

By the way, it is quite weird that DRV_MAX_DEVICES is being used in the
test program, whereas this is the sole business of the driver. To be
addressed later

Signed-off-by: Juan David Gonzalez Cobas <dcobas@cern.ch>