encore: add EX address modifier
This was missing and some modules use this (e.g. SIS3320). Moreover,
these are all the values currently used in the CCDB. It is much
easier to add this in a one-liner modification like here than
cutting and pasting an if statement. This is the point of the previous
commit