encore: restyle code in insencore
Some stylistic issues are addressed here that improve readability.
From deepest to most trivial.

1. Argument checking takes us out by exit, instead of having
nested if's, enhancing readability and cleaning logic
2. Indentation and placement of then clauses adheres now
to a more typical shell programming style

This is an unclean patch, but I did not find a better way to  do
this. Sorry. Next patches will be cleaner