support: use a proper struct * as "global" parameter for lib functions
In the end, passing a handle everywhere is a subtle way to have
a global value passed back and forth. This is bad, but it is even
worse having this passing a void* everywhere, casting unnecessarily
and, worse of all, having redundant new local variable declarations
that serve no purpose and are used (or not) at will (such as it happens
in __vsl_write_reg, among other places copied from it).

Accordingly, no (void *) casts, nor void *'s, are needed anymore