vmeio: use straightforward interface for DMA operations
A function for direct DMA, temporarily named raw_dma_the_right_way
is added, performing the same job as raw_dma but strictly taking
the parameters a DMA transfer requires.

The DMA interface in vmeio has a design flaw: it is associated to
a VME mapping, something completely unrelated to it. This requires
artificially embedding a flag that says whether a mapping will use
DMA or not, which is puzzling. As unrelated information is involved,
the interface for calling it becomes more and more confusing the close
one gets to user space.

This is a first step to correct the situation. A new ioctl() will be
provided as interface for this function. Note that the parameter list
of raw_dma_the_right_way is overlong; probably the vmeio_dma_op
structure is the right thing to pass.