encore: generate lib<module>.[ch] files from database
This is temporarily in dbexctract.py. Now it's kludge and it
cries out for creating a Card class with all the module attributes.

For the time being, this (incorrectly) generates the .c and .h
files. No reference to positional arguments or positional fields
is made, so maintenance is much easier than the one given by
genlib.[ch].awk.