encore: support: remove block offset handling
The implementation of this is completely braindead. This has
to be dealt with at the user library level. The offset is a
property of a block (thus inherited by registers), not a
state of a library at runtime.

I am ANGRY!!!!!!!! This is the worst buggiest kludge EVER!!!!!!
How are we gonna make this work in the user library? By calling
first set_offset, then raw? AAAARRRRRRRGHHHHHHHHHHHH!!!!!