encore: preliminary attempt at a sane libencore interface
Deficiencies in the support library: in the end, vmeio+support
make a huge global area of data sharing. Bad design and bad style.

support leaks many design details of the driver that should be
of no concern to the end user, while hiding others (e.g. DMA acces)
in an ugly manner. Somehow, the user space has to be made aware
of the parameters of mappings, which does not make sense except
for low-level testing facilities. Once again, the whole design (?)
is f***ed up in the interest of the test program, which should be
an afterthought (and easy and straightforward to write provided
a clean interface is there).

This is a first attempt to straighten out this situation. The current
test program will have to live with vmeio_support, because
vmeio+support+test are completely melted into each other. In the
end, a completely separate libencore with a sane design and
a test program based on this, or directly calling vmeio, should
be written instead.