vmeiotest: added register parsing calls to 'driver' command
This patch adds register file parsing capabilities to the
vmeio test shell. All files ending with .regs are searched
in the current directory and the NAME value within the file
checked against the current driver name. If there is a match,
the file is parsed and the list of registers are added.

Signed-off-by: Manohar Vanga <manohar.vanga@cern.ch>