vmeio: provide a decent way to find info about a mapping
The GET_DEVICE interface is a nice way of passing back
and forth global parameters. Bad.

A VMEIO_GET_MAPPING ioctl does the right thing: it gives back
information about one of the mappings. A similar function
should be provided for interrupts.

Doing things the vmeio_get_mapping_s amounts to what DriverGen
did: store everything, even the register descriptions, into
the driver. Which we will not be able to avoid in the end, but
at least we can do in a controlled manner that does not permeate
the design of everything.

Note that this DOES BREAK the test program interface, that
is not very clean either.