vmeio: use words, not bytes as raw operation count units
Using bytes as the size unit for raw/read operations is the unnatural
choice in the application of vmeio to register access (always in
word multiples).

If a wrong (non-word-multiple) byte count is given to raw_read,
it is easy to see that access to unallocated memory occurs.
This modification fixes that bug, by only allowing accesses in
multiples of data_width words, and make interfacing for encore
simpler, because we always read/write registers in multiples
of the data width anyway. Never in bytes.