test: add more stuff to pythono test program
Currently this is a mess. A *true* support library is needed.
Then, we can talk direct user-space calls from Python without
the hassle of ioctl's. Moreover, the problem with the ill-designed
interface for data widths etc. has to be handled more gracefully.

But in the end I got it working.